% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundabouts.R
\docType{data}
\name{roundabouts}
\alias{roundabouts}
\title{Roundabouts in United States, Canada and beyond}
\format{
A data frame with 27887 rows and
13 variables:

\describe{
\item{name}{}
\item{address}{}
\item{coordinates}{}
\item{type}{}
\item{status}{}
\item{year_completed}{}
\item{approaches}{}
\item{driveways}{}
\item{lane_type}{}
\item{functional_class}{}
\item{control_type}{}
\item{other_control_type}{}
\item{previous_control_type}{}
}
}
\source{
\url{https://roundabouts.kittelson.com/}
}
\usage{
roundabouts
}
\description{
A data set containing many of the roundabouts in United States and Canada.
Coordinates and a few characteristics are included with each roundabout.
}
\keyword{datasets}
